+++ src/services/userService.ts
@@ -21,5 +21,11 @@ export const userService = {
 
   async updateUser(username: string, userData: Partial<{ given_name: string; family_name: string; email: string; phone_number: string }>): Promise<ApiResponse> {
     return apiService.put(process.env.REACT_APP_USER_UPDATE_ENDPOINT?.replace('{target_username}', username) || ``, userData);
+  },
+
+  async resetPassword(username: string, newTemporaryPassword: string): Promise<ApiResponse> {
+    return apiService.post(process.env.REACT_APP_USER_RESET_PASSWORD_ENDPOINT?.replace('{target_username}', username) || '', {
+      new_temporary_password: newTemporaryPassword
+    });
   }
 };